% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption_sources_with_dates.r
\name{caption_sources_with_dates}
\alias{caption_sources_with_dates}
\title{Formats captions with dates}
\usage{
caption_sources_with_dates(source, ..., max_width = 180)
}
\arguments{
\item{source}{A source string}

\item{...}{An unknown number of other sources to include}

\item{max_width}{The max_width of the caption}
}
\description{
Formats captions with dates
}
