% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_palettes.R
\name{display_all_palettes}
\alias{display_all_palettes}
\title{Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
\usage{
display_all_palettes()
}
\description{
Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
