% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicurve.R
\name{epicurve.default}
\alias{epicurve.default}
\title{Epicurve}
\usage{
\method{epicurve}{default}(x, granularity_time = "day")
}
\arguments{
\item{granularity_time}{day or isoweek}
}
\description{
Epicurve
}
\examples{
epicurve(x = d_day, granularity_time = "day")
epicurve(x = d_week, granularity_time = "isoyearweek")
}
