% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicurve.R
\name{epicurve}
\alias{epicurve}
\title{Epicurve}
\usage{
epicurve(d, granularity_time = "day")
}
\arguments{
\item{granularity_time}{day or isoweek
epicurve(d = d_day, granularity_time = "day")
epicurve(d = d_week, granularity_time = "isoyearweek")}
}
\description{
Epicurve
}
