% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_nth.R
\name{every_nth}
\alias{every_nth}
\title{every_nth
Useful for plotting every nth discrete value in ggplot2}
\usage{
every_nth(n)
}
\arguments{
\item{n}{nth discrete value}
}
\description{
every_nth
Useful for plotting every nth discrete value in ggplot2
}
\examples{
\dontrun{
scale_x_discrete(NULL, breaks = every_nth(n = 2))
}
}
