% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles_excel.R
\docType{data}
\name{excel_style_background_red}
\alias{excel_style_background_red}
\title{Excel style for red background}
\format{
An object of class \code{Style} of length 1.
}
\usage{
excel_style_background_red
}
\description{
Excel style for red background
}
\keyword{datasets}
