% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles_excel.R
\docType{data}
\name{excel_style_background_yellow}
\alias{excel_style_background_yellow}
\title{Excel style for yellow background}
\format{
An object of class \code{Style} of length 1.
}
\usage{
excel_style_background_yellow
}
\description{
Excel style for yellow background
}
\keyword{datasets}
