% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles_excel.R
\docType{data}
\name{excel_style_num_0}
\alias{excel_style_num_0}
\title{Excel style for number with 0 decimal places}
\format{
An object of class \code{Style} of length 1.
}
\usage{
excel_style_num_0
}
\description{
Excel style for number with 0 decimal places
}
\keyword{datasets}
