% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles_excel.R
\name{excel_style_risk_1_5}
\alias{excel_style_risk_1_5}
\title{Excel style for htmltable cell with a risk scale from 1-5}
\usage{
excel_style_risk_1_5(x, max_risk_value = 5, palette = "blue")
}
\arguments{
\item{x}{Risk value}

\item{max_risk_value}{Max risk value}

\item{palette}{blue/red/green/grayblue/grayred/graygreen}
}
\description{
Excel style for htmltable cell with a risk scale from 1-5
}
