% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles_excel.R
\docType{data}
\name{excel_style_wrap}
\alias{excel_style_wrap}
\title{Excel style wrap}
\format{
An object of class \code{Style} of length 1.
}
\usage{
excel_style_wrap
}
\description{
Excel style wrap
}
\keyword{datasets}
