% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_palettes.R
\name{fhi_pal}
\alias{fhi_pal}
\title{Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
\usage{
fhi_pal(palette = "primary", direction = 1)
}
\arguments{
\item{palette}{as}

\item{direction}{as}
}
\description{
Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
