% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ages.r
\name{format_ages}
\alias{format_ages}
\title{format_ages}
\usage{
format_ages(x)
}
\arguments{
\item{x}{a string (or multiple strings) with an age or an age group}
}
\description{
format_ages
}
\examples{
format_ages("009")
format_ages("005-014")
}
