% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cstidy_age_as.r
\name{format_cstidy_age_as_factor}
\alias{format_cstidy_age_as_factor}
\alias{format_cstidy_age_as_character}
\title{Formats a cstidy age as a factor}
\usage{
format_cstidy_age_as_factor(x)

format_cstidy_age_as_character(x)
}
\arguments{
\item{x}{a string (or multiple strings) with an age or an age group}
}
\description{
Formats a cstidy age as a factor

Formats a cstidy age as a character
}
\examples{
csstyle::format_cstidy_age_as_factor("009")
csstyle::format_cstidy_age_as_factor("005-014")
csstyle::format_cstidy_age_as_character("009")
csstyle::format_cstidy_age_as_character("005-014")
}
