% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{format_date_as_nor}
\alias{format_date_as_nor}
\alias{format_datetime_as_nor}
\alias{format_datetime_as_file}
\title{format_date_nor}
\usage{
format_date_as_nor(x = lubridate::today(), format = "\%d.\%m.\%Y")

format_datetime_as_nor(
  x = lubridate::now(),
  format = "\%d.\%m.\%Y kl. \%H:\%S"
)

format_datetime_as_file(x = lubridate::now(), format = "\%Y_\%m_\%d_\%H\%M\%S")
}
\arguments{
\item{x}{value}

\item{format}{the desired format}
}
\description{
format_date_nor

format_datetime_nor

format_date_nor
}
