% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{format_date_nor}
\alias{format_date_nor}
\title{format_date_nor}
\usage{
format_date_nor(x = lubridate::today(), format = "\%d.\%m.\%Y")
}
\arguments{
\item{x}{value}

\item{format}{the desired format}
}
\description{
format_date_nor
}
