% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{format_datetime_file}
\alias{format_datetime_file}
\title{format_date_nor}
\usage{
format_datetime_file(x = lubridate::now(), format = "\%Y-\%m-\%d_\%H\%M")
}
\arguments{
\item{x}{value}

\item{format}{the desired format}
}
\description{
format_date_nor
}
