% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{format_datetime_nor}
\alias{format_datetime_nor}
\title{format_datetime_nor}
\usage{
format_datetime_nor(x = lubridate::now(), format = "\%d.\%m.\%Y kl. \%H:\%S")
}
\arguments{
\item{x}{value}

\item{format}{the desired format}
}
\description{
format_datetime_nor
}
