% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{format_nor_invlog2_1}
\alias{format_nor_invlog2_1}
\title{format_nor_invlog2_1
Formats as a norwegian number with 1 digit on log-2 scale.
Useful for scale labels}
\usage{
format_nor_invlog2_1(x)
}
\arguments{
\item{x}{value}
}
\description{
format_nor_invlog2_1
Formats as a norwegian number with 1 digit on log-2 scale.
Useful for scale labels
}
