% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{format_nor_num_0}
\alias{format_nor_num_0}
\title{format_nor_num_0
Formats as a norwegian number with 0 digits.
Useful for scale labels}
\usage{
format_nor_num_0(x)
}
\arguments{
\item{x}{value}
}
\description{
format_nor_num_0
Formats as a norwegian number with 0 digits.
Useful for scale labels
}
