% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{format_nor_per100k_2}
\alias{format_nor_per100k_2}
\title{format_nor_num_2
Formats as a norwegian number with 2 digits and the suffix " /100k"
Useful for scale labels}
\usage{
format_nor_per100k_2(x)
}
\arguments{
\item{x}{value}
}
\description{
format_nor_num_2
Formats as a norwegian number with 2 digits and the suffix " /100k"
Useful for scale labels
}
