% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num_as_nor_num_0}
\alias{format_num_as_nor_num_0}
\alias{format_num_as_nor_num_1}
\alias{format_num_as_nor_num_2}
\alias{format_num_as_nor_per100k_0}
\alias{format_num_as_nor_per100k_1}
\alias{format_num_as_nor_per100k_2}
\alias{format_num_as_nor_perc_0}
\alias{format_num_as_nor_perc_1}
\alias{format_num_as_nor_perc_2}
\alias{format_num_as_nor_invlog2_1}
\alias{format_num_as_nor_invlog2_2}
\alias{format_num_as_nor_invlog10_1}
\alias{format_num_as_nor_invlog10_2}
\alias{format_num_as_nor_invloge_1}
\alias{format_num_as_nor_invloge_2}
\title{format_nor_num_0
Formats as a norwegian number with 0 digits.
Useful for scale labels}
\usage{
format_num_as_nor_num_0(x)

format_num_as_nor_num_1(x)

format_num_as_nor_num_2(x)

format_num_as_nor_per100k_0(x)

format_num_as_nor_per100k_1(x)

format_num_as_nor_per100k_2(x)

format_num_as_nor_perc_0(x)

format_num_as_nor_perc_1(x)

format_num_as_nor_perc_2(x)

format_num_as_nor_invlog2_1(x)

format_num_as_nor_invlog2_2(x)

format_num_as_nor_invlog10_1(x)

format_num_as_nor_invlog10_2(x)

format_num_as_nor_invloge_1(x)

format_num_as_nor_invloge_2(x)
}
\arguments{
\item{x}{value}
}
\description{
format_nor_num_0
Formats as a norwegian number with 0 digits.
Useful for scale labels

format_nor_num_1
Formats as a norwegian number with 1 digits.
Useful for scale labels

format_nor_num_2
Formats as a norwegian number with 2 digits.
Useful for scale labels

format_nor_num_0
Formats as a norwegian number with 0 digits and the suffix " /100k".
Useful for scale labels

format_nor_num_1
Formats as a norwegian number with 1 digits and the suffix " /100k"
Useful for scale labels

format_nor_num_2
Formats as a norwegian number with 2 digits and the suffix " /100k"
Useful for scale labels

format_nor_perc_0
Formats as a norwegian number with 0 digits and puts a % sign afterwards.
Useful for scale labels

format_nor_perc_1
Formats as a norwegian number with 1 digits and puts a % sign afterwards.
Useful for scale labels

format_nor_perc_2
Formats as a norwegian number with 2 digits and puts a % sign afterwards.
Useful for scale labels

format_nor_invlog2_1
Formats as a norwegian number with 1 digit on log-2 scale.
Useful for scale labels

format_nor_invlog2_2
Formats as a norwegian number with 2 digits on log-2 scale.
Useful for scale labels

format_nor_invlog10_1
Formats as a norwegian number with 1 digit on log-10 scale.
Useful for scale labels

format_nor_invlog10_2
Formats as a norwegian number with 2 digits on log-10 scale.
Useful for scale labels

format_nor_invloge_1
Formats as a norwegian number with 1 digit on log scale.
Useful for scale labels

format_nor_invloge_2
Formats as a norwegian number with 2 digits on log scale.
Useful for scale labels
}
