% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num_as_nor_invlog10_1}
\alias{format_num_as_nor_invlog10_1}
\alias{format_num_as_nor_invlog10_2}
\title{Formats as a norwegian number with 1 digit on log-10 scale.
Useful for scale labels}
\usage{
format_num_as_nor_invlog10_1(x)

format_num_as_nor_invlog10_2(x)
}
\arguments{
\item{x}{value}
}
\value{
Character vector

Character vector
}
\description{
Formats as a norwegian number with 1 digit on log-10 scale.
Useful for scale labels

Formats as a norwegian number with 2 digits on log-10 scale.
Useful for scale labels
}
