% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num_as_nor_num_0}
\alias{format_num_as_nor_num_0}
\alias{format_num_as_nor_num_1}
\alias{format_num_as_nor_num_2}
\title{format_nor_num_0
Formats as a norwegian number with 0 digits.
Useful for scale labels}
\usage{
format_num_as_nor_num_0(x)

format_num_as_nor_num_1(x)

format_num_as_nor_num_2(x)
}
\arguments{
\item{x}{value}
}
\description{
format_nor_num_0
Formats as a norwegian number with 0 digits.
Useful for scale labels

format_nor_num_1
Formats as a norwegian number with 1 digits.
Useful for scale labels

format_nor_num_2
Formats as a norwegian number with 2 digits.
Useful for scale labels
}
