% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_stepribbon.R
\name{geom_stepribbon}
\alias{geom_stepribbon}
\title{stat_stepribbon
An extension of ggplot2's geom_ribbon}
\usage{
geom_stepribbon(
  mapping = NULL,
  data = NULL,
  geom = "ribbon",
  position = "identity",
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Mapping}

\item{data}{Data}

\item{geom}{Geom}

\item{position}{Position}

\item{inherit.aes}{Inherit.aes}

\item{...}{Dots}
}
\description{
stat_stepribbon
An extension of ggplot2's geom_ribbon
}
