% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltable_css.R
\docType{data}
\name{htmltable_css_background_yellow}
\alias{htmltable_css_background_yellow}
\title{CSS style for htmltable for yellow background}
\format{
An object of class \code{character} of length 1.
}
\usage{
htmltable_css_background_yellow
}
\description{
CSS style for htmltable for yellow background
}
\keyword{datasets}
