% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltable_css.R
\name{htmltable_css_header}
\alias{htmltable_css_header}
\title{CSS for htmltable header}
\usage{
htmltable_css_header(font_size = 16, border_top = "2px solid grey")
}
\arguments{
\item{font_size}{font_size}

\item{border_top}{Border top style}
}
\description{
CSS for htmltable header
}
