% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltable_css.R
\name{htmltable_css_risk_1_5}
\alias{htmltable_css_risk_1_5}
\title{CSS for htmltable cell with a risk scale from 1-5}
\usage{
htmltable_css_risk_1_5(x, font_size = 14, max_risk_value = 5, palette = "blue")
}
\arguments{
\item{x}{Risk value}

\item{font_size}{font_size}

\item{max_risk_value}{Max risk value}

\item{palette}{blue/red/green/grayblue/grayred/graygreen}
}
\description{
CSS for htmltable cell with a risk scale from 1-5
}
