% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltable_css.R
\name{htmltable_css_style_matrix}
\alias{htmltable_css_style_matrix}
\title{Empty CSS style matrix}
\usage{
htmltable_css_style_matrix(tab)
}
\arguments{
\item{tab}{The table that gives the dimensions}
}
\description{
Empty CSS style matrix
}
