% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable.R
\name{huxtable_theme_fhi_basic}
\alias{huxtable_theme_fhi_basic}
\title{huxtable_theme_fhi_basic}
\usage{
huxtable_theme_fhi_basic(ht, position = "center")
}
\arguments{
\item{ht}{A huxtable}

\item{position}{Position}
}
\description{
huxtable_theme_fhi_basic
}
