% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_code_to_factor.r
\name{location_code_to_factor}
\alias{location_code_to_factor}
\title{Makes a factor with a label from a vector of location codes}
\usage{
location_code_to_factor(
  x,
  label = splstyle::config$location_code_to_factor_label,
  label_if_not_unique = splstyle::config$location_code_to_factor_label_if_not_unique,
  reference = fhidata::norway_locations_names(),
  direction = 1
)
}
\arguments{
\item{x}{A vector of location codes}

\item{label}{The label of the elements in the new factor}

\item{label_if_not_unique}{The label of elements that are not unique}

\item{reference}{Where you get your data}

\item{direction}{-1 is reverse}
}
\description{
Makes a factor with a label from a vector of location codes
}
