% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_mem.R
\name{make_influenza_threshold_chart}
\alias{make_influenza_threshold_chart}
\title{A treshold chart of the type used for influenza}
\usage{
make_influenza_threshold_chart(
  data,
  title,
  lang = "NB",
  weeks = c(40, 20),
  color_palette = "fhi",
  legend_control = "legend"
)
}
\arguments{
\item{data}{A data.table for a season of influenza}

\item{title}{the title for the chart}

\item{lang}{language of labels}

\item{weeks}{start and end week for the plot}

\item{color_palette}{which color palette to use (fhi uses standard splstyle, otherwise influensa palette is used)}

\item{legend_control}{using a legend or text on the chart}
}
\description{
A treshold chart of the type used for influenza
}
