% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_line_threshold.R
\name{make_line_brush_plot}
\alias{make_line_brush_plot}
\title{MakeLineBrushPlot}
\usage{
make_line_brush_plot(pd, x, dataVal, L2, L3, GetCols)
}
\arguments{
\item{pd}{a}

\item{x}{a}

\item{dataVal}{a}

\item{L2}{a}

\item{L3}{a}

\item{GetCols}{a}
}
\description{
MakeLineBrushPlot
}
