% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_line_excess.R
\name{make_line_excess_graph}
\alias{make_line_excess_graph}
\title{make_line_excess_graph}
\usage{
make_line_excess_graph(
  pd,
  x,
  dataVal,
  dataZ,
  dataCIL = NULL,
  dataCIU = NULL,
  allPoints = TRUE,
  title = NULL,
  pointShift = 0,
  xShift = 0,
  weekNumbers = FALSE,
  step = FALSE,
  GetCols,
  legend_position = "right"
)
}
\arguments{
\item{pd}{a}

\item{x}{a}

\item{dataVal}{a}

\item{dataZ}{a}

\item{dataCIL}{a}

\item{dataCIU}{a}

\item{allPoints}{a}

\item{title}{a}

\item{pointShift}{a}

\item{xShift}{a}

\item{weekNumbers}{a}

\item{step}{a}

\item{GetCols}{a}

\item{legend_position}{a}
}
\description{
make_line_excess_graph
}
