% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_line_threshold.R
\name{make_line_threshold_plot}
\alias{make_line_threshold_plot}
\title{Creates a line threshold plot}
\usage{
make_line_threshold_plot(
  pd,
  x,
  dataVal,
  dataCIL = NULL,
  dataCIU = NULL,
  L1,
  L2,
  L3,
  L4,
  allPoints = TRUE,
  title = NULL,
  pointShift = 0,
  xShift = 0,
  weekNumbers = FALSE,
  step = FALSE,
  GetCols,
  legend_position = "right"
)
}
\arguments{
\item{pd}{A data.frame}

\item{x}{Name of the column in `pd` that is the `x` value}

\item{dataVal}{Name of the column in `pd` that is the `y` value}

\item{dataCIL}{Name of the column in `pd` that is the value of the lower confidence interval of the `y` value}

\item{dataCIU}{Name of the column in `pd` that is the value of the upper confidence interval of the `y` value}

\item{L1}{Name of the column in `pd` that is the value of the first threshold}

\item{L2}{Name of the column in `pd` that is the value of the second threshold}

\item{L3}{Name of the column in `pd` that is the value of the third threshold}

\item{L4}{Name of the column in `pd` that is the value of the fourth threshold}

\item{allPoints}{Should all points be displayed on the line?}

\item{title}{Title of the graph}

\item{pointShift}{How much will the points be shifted?}

\item{xShift}{How much will the x values be shifted?}

\item{weekNumbers}{Display week numbers?}

\item{step}{Step graph or normal line graph?}

\item{GetCols}{Vector of colours}

\item{legend_position}{a}
}
\description{
Creates a line threshold plot
}
