% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal_timeseries.r
\name{plot_seasonal_timeseries}
\alias{plot_seasonal_timeseries}
\alias{plot_seasonal_timeseries.default}
\title{plot_seasonal_timeseries}
\usage{
plot_seasonal_timeseries(x, ...)

\method{plot_seasonal_timeseries}{default}(
  x,
  var_x = "isoyearweek",
  var_y,
  breaks_x = seq(1, 52, 4),
  breaks_y = csstyle::pretty_breaks(6),
  lab_main = NULL,
  lab_sub = NULL,
  lab_caption = "hi",
  lab_date = "\%Y-\%m-\%d",
  lab_y = NULL,
  lab_x = NULL,
  lab_legend = NULL,
  legend_position = "bottom",
  legend_direction = "horizontal",
  format_y = csstyle::format_num_as_nor_num_0,
  facet_wrap = NULL,
  facet_ncol = NULL,
  palette = "primary",
  palette_dir = 1,
  scale_y = "free",
  base_size = 12,
  geom_point = FALSE,
  fn_convert_isoyearweek_to_season = cstime::isoyearweek_to_season_c,
  fn_convert_isoyearweek_to_seasonweek = cstime::isoyearweek_to_seasonweek_n,
  fn_convert_seasonweek_to_isoweek = cstime::seasonweek_to_isoweek_c,
  fn_convert_isoweek_to_seasonweek = cstime::isoweek_to_seasonweek_n,
  ...
)
}
\arguments{
\item{x}{Dataset}

\item{...}{Not currently used.}

\item{var_x}{"date" or "isoyearweek"}

\item{var_y}{The name of the variable to use on the y-axis of the graph}

\item{breaks_x}{Use csstyle::every_nth() to choose how many ticks to show on the x-axis}

\item{breaks_y}{Use csstyle::pretty_breaks() to add ticks on the y-axis}

\item{lab_main}{The main title of the graph}

\item{lab_sub}{The subtitle of the graph}

\item{lab_caption}{If not specified, csstyle::fhi_caption() is used as the lab_caption.}

\item{lab_date}{How the dates on the x-axis should be formatted if var_x = "date"}

\item{lab_y}{The label of the y-axis}

\item{lab_x}{The label of the x-axis}

\item{lab_legend}{The label of the legend.}

\item{legend_position}{The position the legend should have. If not specified, "bottom" is used.}

\item{legend_direction}{layout of items in legend ("horizontal" or "vertical")}

\item{format_y}{How the y-axis ticks should be formatted. For example csstyle::format_nor_num_0 or csstyle::format_nor_perc_0}

\item{facet_wrap}{What column in the dataset to use to split the dataset.}

\item{facet_ncol}{How many columns with graphs if facet_wrap is used.}

\item{palette}{What palette to use for the lines. The default is "primary".}

\item{palette_dir}{1 or -1.}

\item{scale_y}{How to scale the y-axis if the graph is split with facet_wrap. Free or fixed.}

\item{base_size}{The base size of the plot.}

\item{geom_point}{TRUE if points should be included in the graph.}

\item{fn_convert_isoyearweek_to_season}{A function that converts from isoyearweek to season}

\item{fn_convert_isoyearweek_to_seasonweek}{A function that converts from isoyearweek to seasonweek}

\item{fn_convert_seasonweek_to_isoweek}{A function that converts from seasonweek to isoweek}

\item{fn_convert_isoweek_to_seasonweek}{A function that convets from isoweek to seasonweek}
}
\description{
If the dataset is already long it needs to include the following columns: variable, name_outcome and n.
}
