% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.r
\name{plot_timeseries.default}
\alias{plot_timeseries.default}
\title{plot_timeseries}
\usage{
\method{plot_timeseries}{default}(
  data,
  var_x = "isoyearweek",
  var_y,
  breaks_x = NULL,
  breaks_y = NULL,
  lab_main = NULL,
  lab_sub = NULL,
  lab_caption = fhi_caption(),
  lab_y = NULL,
  lab_x = NULL,
  lab_legend = NULL,
  legend_position = "bottom",
  legend_direction = "horizontal",
  format_y = format_nor_num_0,
  facet_wrap = NULL,
  facet_ncol = NULL,
  palette = "primary",
  palette_dir = 1,
  scale_y = "free",
  base_size = 12,
  wide_table = TRUE
)
}
\arguments{
\item{data}{Dataset}

\item{var_x}{"date" or "isoyearweek"}

\item{var_y}{The name of the variable to use on the y-axis of the graph}

\item{breaks_x}{Use splstyle::every_nth() to choose how many ticks to show on the x-axis}

\item{breaks_y}{X}

\item{lab_main}{The main title of the graph}

\item{lab_sub}{The subtitle of the graph}

\item{lab_caption}{If not specified, splstyle::fhi_caption() is used as the lab_caption.}

\item{lab_y}{The label of the y-axis}

\item{lab_x}{The label of the x-axis}

\item{lab_legend}{The label of the legend.}

\item{legend_position}{The position the legend should have. If not specified, "bottom" is used.}

\item{legend_direction}{X}

\item{format_y}{How the y-axis ticks should be formatted. For example splstyle::format_nor_num_0 or fhiplot::format_nor_perc_0}

\item{facet_wrap}{What column in the dataset to use to split the dataset.}

\item{facet_ncol}{How many columns with graphs if facet_wrap is used.}

\item{palette}{What palette to use for the lines. The default is "primary".}

\item{palette_dir}{1 or -1.}

\item{scale_y}{How to scale the y-axis if the graph is split with facet_wrap. Free or fixed.}

\item{base_size}{The base size of the plot.}

\item{wide_table}{X}
}
\description{
plot_timeseries
}
\examples{
plot_timeseries(norway_covid19_cases_by_time_location[granularity_geo == "nation" & granularity_time == "isoweek"], var_y = c("Covid cases" = "covid19_cases_testdate_n"), breaks_x = every_nth(8), breaks_y = splstyle::pretty_breaks(5))
plot_timeseries(norway_covid19_cases_by_time_location[granularity_geo == "county" & granularity_time == "isoweek"], var_y = c("Covid cases" = "covid19_cases_testdate_n"), breaks_x = every_nth(8), breaks_y = splstyle::pretty_breaks(5), facet_wrap = "location_code")
}
