% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.r
\name{plot_timeseries}
\alias{plot_timeseries}
\title{plot_timeseries}
\usage{
plot_timeseries(
  data,
  var_x = "isoyearweek",
  var_y,
  breaks_x = NULL,
  lab_main = NULL,
  lab_sub = NULL,
  lab_caption = fhi_caption(),
  lab_y = NULL,
  lab_x = NULL,
  lab_legend = NULL,
  legend_position = "bottom",
  format_y = format_nor_num_0,
  facet_wrap = NULL,
  facet_ncol = NULL,
  palette = "primary",
  palette_dir = 1,
  scale_y = "free"
)
}
\arguments{
\item{data}{Dataset}

\item{var_x}{"date" or "isoyearweek"}

\item{var_y}{The name of the variable to use on the y-axis of the graph}

\item{breaks_x}{Use splstyle::every_nth() to choose how many ticks to show on the x-axis}

\item{lab_main}{The main title of the graph}

\item{lab_sub}{The subtitle of the graph}

\item{lab_caption}{If not specified, splstyle::fhi_caption() is used as the lab_caption.}

\item{lab_y}{The label of the y-axis}

\item{lab_x}{The label of the x-axis}

\item{lab_legend}{The label of the legend.}

\item{legend_position}{The position the legend should have. If not specified, "bottom" is used.}

\item{format_y}{How the y-axis ticks should be formatted. For example splstyle::format_nor_num_0 or fhiplot::format_nor_perc_0}

\item{facet_wrap}{What column in the dataset to use to split the dataset.}

\item{facet_ncol}{How many columns with graphs if facet_wrap is used.}

\item{palette}{What palette to use for the lines. The default is "primary".}

\item{palette_dir}{1 or -1.}

\item{scale_y}{How to scale the y-axis if the graph is split with facet_wrap. Free or fixed.}
}
\description{
plot_timeseries
}
