% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{pretty_breaks}
\alias{pretty_breaks}
\title{pretty_breaks}
\usage{
pretty_breaks(n = 5, digits = 0, break_with_four_digits = T, ...)
}
\arguments{
\item{n}{a}

\item{digits}{number of decimal places}

\item{break_with_four_digits}{Whether break with four digits. Default is TRUE (optional)}

\item{...}{dots}
}
\description{
pretty_breaks
}
