% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_code.R
\name{print.github_code}
\alias{print.github_code}
\title{Print github_code}
\usage{
\method{print}{github_code}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{Further arguments}
}
\description{
Print github_code
}
