% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_palettes.R
\name{scale_color_fhi}
\alias{scale_color_fhi}
\title{Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
\usage{
scale_color_fhi(..., palette = "primary", direction = 1)
}
\arguments{
\item{...}{as}

\item{palette}{as}

\item{direction}{as}
}
\description{
Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
