% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_palettes.R
\name{scale_fill_cs}
\alias{scale_fill_cs}
\title{Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
\usage{
scale_fill_cs(..., palette = "primary", direction = 1)
}
\arguments{
\item{...}{a}

\item{palette}{a}

\item{direction}{a}
}
\description{
Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
