% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_environments.r
\name{set_config}
\alias{set_config}
\title{set_config}
\usage{
set_config(
  location_code_to_factor_label = "location_name",
  location_code_to_factor_label_if_not_unique = "location_name_description_nb"
)
}
\arguments{
\item{location_code_to_factor_label}{The label}

\item{location_code_to_factor_label_if_not_unique}{The label if something is not unique with the first laben}
}
\description{
set_config
}
