% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_theme.R
\name{theme_cs}
\alias{theme_cs}
\alias{set_x_axis_vertical}
\title{ggplot2 theme}
\usage{
theme_cs(
  base_size = 16,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  legend_position = "right",
  x_axis_vertical = FALSE,
  panel_on_top = TRUE,
  panel.grid.major.x = FALSE,
  panel.grid.minor.x = FALSE,
  panel.grid.major.y = TRUE,
  panel.grid.minor.y = TRUE
)

set_x_axis_vertical()
}
\arguments{
\item{base_size}{a}

\item{base_family}{a}

\item{base_line_size}{a}

\item{base_rect_size}{a}

\item{legend_position}{a}

\item{x_axis_vertical}{FALSE}

\item{panel_on_top}{a}

\item{panel.grid.major.x}{Set to FALSE}

\item{panel.grid.minor.x}{Set to FALSE}

\item{panel.grid.major.y}{Set to FALSE}

\item{panel.grid.minor.y}{Set to FALSE}
}
\value{
ggplot2 theme

ggplot2 theme component
}
\description{
ggplot2 theme

Sets the x axis to be vertical
}
