% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_fhi_basic}
\alias{theme_fhi_basic}
\title{theme_fhi_basic}
\usage{
theme_fhi_basic(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  legend_position = "right"
)
}
\arguments{
\item{base_size}{a}

\item{base_family}{a}

\item{base_line_size}{a}

\item{base_rect_size}{a}

\item{legend_position}{a}
}
\description{
theme_fhi_basic
}
