% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_fhi_lines_horizontal}
\alias{theme_fhi_lines_horizontal}
\title{theme_fhi_lines_horizontal}
\usage{
theme_fhi_lines_horizontal(
  base_size = 12,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  legend_position = "right",
  panel_on_top = TRUE,
  panel.grid.major.x = element_blank(),
  panel.grid.minor.x = element_blank(),
  panel.grid.major.y = element_line(colour = "black", size = rel(0.1)),
  panel.grid.minor.y = element_line(colour = "black", size = rel(0.05))
)
}
\arguments{
\item{base_size}{a}

\item{base_family}{a}

\item{base_line_size}{a}

\item{base_rect_size}{a}

\item{legend_position}{a}

\item{panel_on_top}{a}

\item{panel.grid.major.x}{Set to element_blank() to remove}

\item{panel.grid.minor.x}{Set to element_blank() to remove}

\item{panel.grid.major.y}{Set to element_blank() to remove}

\item{panel.grid.minor.y}{Set to element_blank() to remove}
}
\description{
theme_fhi_lines_horizontal
}
