% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_environments.r
\docType{data}
\name{warning_color}
\alias{warning_color}
\title{Warning color}
\format{
An object of class \code{character} of length 3.
}
\usage{
warning_color
}
\description{
Warning color
}
\keyword{datasets}
