% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_code.R
\name{as_github_code}
\alias{as_github_code}
\title{An easy way to print github code}
\usage{
as_github_code(url)
}
\arguments{
\item{url}{URL from github}
}
\value{
Class github_code
}
\description{
An easy way to print github code
}
\examples{
\dontrun{
x <- as_github_code("https://github.com/x/y/blob/main/file.txt")
print(x, lines = 3:5)
}
}
