% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{format_date_as_nor}
\alias{format_date_as_nor}
\alias{format_datetime_as_nor}
\alias{format_datetime_as_file}
\title{Format date in the Norwegian format}
\usage{
format_date_as_nor(x = lubridate::today(), format = "\%d.\%m.\%Y")

format_datetime_as_nor(
  x = lubridate::now(),
  format = "\%d.\%m.\%Y kl. \%H:\%S"
)

format_datetime_as_file(x = lubridate::now(), format = "\%Y_\%m_\%d_\%H\%M\%S")
}
\arguments{
\item{x}{Datetime}

\item{format}{The desired format}
}
\value{
A character vector

A character vector

A character vector
}
\description{
Format date in the Norwegian format

Format datetime in the Norwegian format

Format datetime in a file format
}
