% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_location_code_as_.r
\name{format_location_code_as_factor}
\alias{format_location_code_as_factor}
\alias{format_location_code_as_character}
\title{Makes a factor with a label from a vector of location codes}
\usage{
format_location_code_as_factor(
  x,
  label = NULL,
  label_if_not_unique = NULL,
  reference = csdata::nor_locations_names(),
  direction = 1
)

format_location_code_as_character(
  x,
  label = NULL,
  label_if_not_unique = NULL,
  reference = csdata::nor_locations_names(),
  direction = 1
)
}
\arguments{
\item{x}{A vector of location codes}

\item{label}{The label you want the elements in the character to have}

\item{label_if_not_unique}{The label of elements that are not unique}

\item{reference}{Where you get your data}

\item{direction}{-1 is reverse}
}
\description{
Makes a factor with a label from a vector of location codes

Makes a character from a vector of location codes
}
