% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num_as_nor_perc_0}
\alias{format_num_as_nor_perc_0}
\alias{format_num_as_nor_perc_1}
\alias{format_num_as_nor_perc_2}
\title{Formats as a norwegian number with 0 digits and puts a % sign afterwards.
Useful for scale labels}
\usage{
format_num_as_nor_perc_0(x)

format_num_as_nor_perc_1(x)

format_num_as_nor_perc_2(x)
}
\arguments{
\item{x}{value}
}
\value{
Character vector

Character vector

Character vector
}
\description{
Formats as a norwegian number with 0 digits and puts a % sign afterwards.
Useful for scale labels

Formats as a norwegian number with 1 digits and puts a % sign afterwards.
Useful for scale labels

Formats as a norwegian number with 2 digits and puts a % sign afterwards.
Useful for scale labels
}
