% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epicurve.R
\name{plot_epicurve}
\alias{plot_epicurve}
\alias{plot_epicurve.default}
\title{Epicurve}
\usage{
plot_epicurve(x, ...)

\method{plot_epicurve}{default}(
  x,
  type = "single",
  fill_var = NULL,
  fill_lab = NULL,
  facet_wrap = NULL,
  facet_ncol = NULL,
  var_x,
  var_y,
  breaks_x = ggplot2::waiver(),
  breaks_y = csstyle::pretty_breaks(6),
  lab_x = NULL,
  lab_y = NULL,
  lab_main = NULL,
  lab_sub = NULL,
  lab_caption = "hi",
  lab_date = "\%Y-\%m-\%d",
  format_y = csstyle::format_num_as_nor_num_0,
  scale_y = "free",
  palette = "primary",
  base_size = 12,
  ...
)
}
\arguments{
\item{x}{Dataset}

\item{...}{Not currently used.}

\item{type}{"single", "stacked" or "dodged"}

\item{fill_var}{variable to fill by.}

\item{fill_lab}{fill label}

\item{facet_wrap}{What column in the dataset to use to split the dataset.}

\item{facet_ncol}{How many columns with graphs}

\item{var_x}{"date" or "isoyearweek"}

\item{var_y}{The name of the variable to use on the y-axis of the graph}

\item{breaks_x}{Use csstyle::every_nth() to choose how many ticks to show on the x-axis}

\item{lab_x}{The label of the x-axis}

\item{lab_y}{The label of the y-axis}

\item{lab_main}{The main title of the graph}

\item{lab_sub}{The subtitle of the graph}

\item{lab_caption}{If not specified, csstyle::fhi_caption() is used as the lab_caption.}

\item{lab_date}{How the dates on the x-axis should be formatted if var_x = "date"}

\item{format_y}{How the y-axis ticks should be formatted. For example csstyle::format_nor_num_0 or csstyle::format_nor_perc_0}

\item{scale_y}{How to scale the y-axis if the graph is split with facet_wrap. Free or fixed.}

\item{palette}{what palette to use}

\item{base_size}{size of plot}
}
\description{
Epicurve

Epicurve
}
\examples{
csstyle::plot_epicurve(
  covidnor::total[
    granularity_geo=="nation" &
    granularity_time=="day" &
    isoyear==2020
  ],
  type = "single",
  var_x = "date",
  var_y = "hospital_admissions_main_cause_n"
)
csstyle::plot_epicurve(
  covidnor::total[
    granularity_geo=="nation" &
    granularity_time=="isoweek" &
    isoyear==2020
  ],
  type = "single",
  var_x = "isoyearweek",
  var_y = "hospital_admissions_main_cause_n"
)
csstyle::plot_epicurve(
  covidnor::total[
    granularity_geo=="county" &
    granularity_time=="day" &
    isoyear==2020
  ],
  type = "stacked",
  var_x = "date",
  fill_var = "location_name",
  var_y = "cases_by_testdate_n"
)
csstyle::plot_epicurve(
  covidnor::total[
    granularity_geo=="county" &
    granularity_time=="isoweek" &
    isoyear==2020
  ],
  type = "stacked",
  var_x = "isoyearweek",
  fill_var = "location_name",
  var_y = "cases_by_testdate_n"
)
csstyle::plot_epicurve(
  covidnor::total[
    location_code \%in\% c("county_nor34", "county_nor38", "county_nor11") &
    granularity_time=="day" &
    isoyear==2020
  ],
  type = "dodged",
  fill_var = "location_name",
  var_x = "date",
  var_y = "cases_by_testdate_n"
)
csstyle::plot_epicurve(
  covidnor::total[
    location_code \%in\% c("county_nor34", "county_nor38", "county_nor11") &
    granularity_time=="isoweek" &
    isoyear==2020
  ],
  type = "dodged",
  fill_var = "location_name",
  var_x = "isoyearweek",
  var_y = "cases_by_testdate_n"
)
}
