% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{pretty_breaks}
\alias{pretty_breaks}
\title{Pretty breaks}
\usage{
pretty_breaks(n = 5, formatting_fn = csstyle::format_num_as_nor_num_0, ...)
}
\arguments{
\item{n}{a}

\item{formatting_fn}{Formatting function}

\item{...}{dots}
}
\value{
A function that can be used for breaks on graphing.
}
\description{
Pretty breaks
}
