% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_a4.R
\name{save_a4}
\alias{save_a4}
\title{Save ggplot in A4 scale}
\usage{
save_a4(q, filename, landscape = T, scaling_factor = 1)
}
\arguments{
\item{q}{Plot}

\item{filename}{filename}

\item{landscape}{landscape dimensions?}

\item{scaling_factor}{How much larger/smaller than A4?}
}
\value{
Nothing.
}
\description{
Save ggplot in A4 scale
}
